separate client and services definition
