+++ packages/aws-client/client.js
@@ -1,0 +1,44 @@
+const AWS = require('aws-sdk');
+const { noop } = require('@cumulus/common/util');
+
+const { inTestMode, testAwsClient } = require('./test-utils');
+
+exports.region = process.env.AWS_DEFAULT_REGION || 'us-east-1';
+AWS.config.update({ region: exports.region });
+
+// Workaround upload hangs. See: https://github.com/andrewrk/node-s3-client/issues/74'
+AWS.util.update(AWS.S3.prototype, { addExpect100Continue: noop });
+AWS.config.setPromisesDependency(Promise);
+
+const memoize = (fn) => {
+  let memo = null;
+  return (options) => {
+    if (!memo) memo = fn(options);
+    return memo;
+  };
+};
+
+/**
+ * Return a function which, when called, will return an AWS service object
+ *
+ * Note: The returned service objects are cached, so there will only be one
+ *       instance of each service object per process.
+ *
+ * @param {Function} Service - an AWS service object constructor function
+ * @param {string} version - the API version to use
+ * @returns {Function} - a function which, when called, will return an AWS service object
+ */
+const awsClient = (Service, version = null) => {
+  const options = {};
+  if (version) options.apiVersion = version;
+
+  if (inTestMode()) {
+    if (AWS.DynamoDB.DocumentClient.serviceIdentifier === undefined) {
+      AWS.DynamoDB.DocumentClient.serviceIdentifier = 'dynamodb';
+    }
+    return memoize((o) => testAwsClient(Service, Object.assign(options, o)));
+  }
+  return memoize((o) => new Service(Object.assign(options, o)));
+};
+
+module.exports = awsClient;
